1. 라이브러리 프로젝트(WanyCore) 추가. 2. 라이브러리 테스트용 프로젝트 템플릿 및  프로젝트 항목별로 추가.

+++ WanyCore/DXDevice.cpp
@@ -136,8 +136,8 @@ void DXDevice::createViewPort()
 	// 5) ºä Æ÷Æ® ¼³Á¤
 	//////////////////////////////////////////////////////////////////////
 	D3D11_VIEWPORT vp;
-	vp.Width = clientRect.right;
-	vp.Height = clientRect.bottom;
+	vp.Width = static_cast<float>(clientRect.right);
+	vp.Height = static_cast<float>(clientRect.bottom);
 	vp.TopLeftX = 0;
 	vp.TopLeftY = 0;
 	vp.MaxDepth = 1.0f;
